Use useGetContext in form state to have automatic types, rename FormStateProvider -> useFormState
