+++ next/components/forms/FormPage.tsx
@@ -3,13 +3,13 @@ import MenuList from 'components/forms/steps/MenuList'
 
 import { defaultFormStateBehavior, rjfsValidator } from '../../frontend/utils/form'
 import FormControls from './FormControls'
-import { useFormState } from './FormStateProvider'
 import FormUploadXml from './FormUploadXml'
 import FormModals from './segments/FormModals/FormModals'
 import FormHeader from './simple-components/FormHeader'
 import StepperView from './steps/StepperView'
 import FormSummary from './steps/Summary/FormSummary'
 import ThemedForm from './ThemedForm'
+import { useFormState } from './useFormState'
 
 const FormPage = () => {
   const {
