+++ next/components/forms/FormPageWrapper.tsx
@@ -8,11 +8,11 @@ import AccountPageLayout from '../layouts/AccountPageLayout'
 import PageWrapper from '../layouts/PageWrapper'
 import { GetSSRCurrentAuth } from '../logic/ServerSideAuthProvider'
 import FormPage from './FormPage'
-import { FormStateProvider } from './FormStateProvider'
 import ThankYouFormSection from './segments/AccountSections/ThankYouSection/ThankYouFormSection'
 import { FormFileUploadStateProvider } from './useFormFileUpload'
 import { FormModalsProvider } from './useFormModals'
 import { FormSentRenderer } from './useFormSent'
+import { FormStateProvider } from './useFormState'
 
 export type FormPageWrapperProps = {
   schema: RJSFSchema
