+++ next/components/forms/segments/FormModals/FormModals.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'next-i18next'
 import React from 'react'
 
 import { useFormExportImport } from '../../../../frontend/hooks/useFormExportImport'
-import { useFormState } from '../../FormStateProvider'
 import Button from '../../simple-components/ButtonNew'
 import { useFormModals } from '../../useFormModals'
+import { useFormState } from '../../useFormState'
 import MessageModal, { MessageModalProps } from '../../widget-components/Modals/MessageModal'
 import IdentityVerificationModal from '../IdentityVerificationModal/IdentityVerificationModal'
 import RegistrationModal from '../RegistrationModal/RegistrationModal'
