+++ next/frontend/hooks/useFormExportImport.tsx
@@ -5,9 +5,9 @@ import { AxiosResponse } from 'axios'
 import { useTranslation } from 'next-i18next'
 import React, { createContext, PropsWithChildren, useContext, useRef } from 'react'
 
-import { useFormState } from '../../components/forms/FormStateProvider'
 import { RegistrationModalType } from '../../components/forms/segments/RegistrationModal/RegistrationModal'
 import { useFormModals } from '../../components/forms/useFormModals'
+import { useFormState } from '../../components/forms/useFormState'
 import { readFileToString } from '../utils/file'
 import { downloadBlob } from '../utils/general'
 import { useServerSideAuth } from './useServerSideAuth'
